Add nvidia_tegra_drivers_support interface
This interface will give snaps access to the iGPU on nvidia tegra
platforms. The included udev rules will tag the device nodes related to
the Nvidia dma fence, the nvmap memory manager, the non-debug iGPU and
the renderer.

The apparmor profiles and seccomp rules were determined iteratively
using snappy-debug and testing it with three different snaps that were
created. One for the nvidia-tegra-drivers-36 package which also contains
the `nvidia-smi` tool, one for cuda-samples[2] and one for
libnvinfer-samples[3] in nvidia's jetson/common package repository which
contains TensorRT samples.

[1] https://launchpad.net/~ubuntu-tegra/+archive/ubuntu/updates
[2] https://github.com/NVIDIA/cuda-samples
[3] https://repo.download.nvidia.com/jetson#Jetpack%206.0

Signed-off-by: Sebastian Weyer <sebastian.weyer@canonical.com>