i/prompting/patterns: compute precedence based on matching path
With knowledge of the path which matched every pattern variant we're
comparing, we can compute which of those variants has the highest
precedence.

For example, the precedence of the following pattern variants cannot be
computed a priori:

- /foo/**/fizz/**
- /foo/**/buzz/**

However, with the matching path "/foo/bar/fizz/buzz", we can say that
"/foo/**/fizz/**" has precedence, since "fizz" matches earlier in the
path than "buzz". If instead the path were "/foo/bar/buzz/fizz", then
"/foo/**/buzz/**" would have precedence, for the same reason.

To compute this, we add a regular expression snippet for every
component, concatenate and compile them into a single regex for the
pattern variant, and match it against the path when computing
precedence, keeping track of submatches. As we iterate through the
components of the pattern variant, we also iterate through the
submatches, and we are thus able to tell more about how many bytes each
component matched. For variable-length component types (e.g. globstars
and doublestars), matching fewer bytes yields higher precedence.
Conversely, for literals, matching more bytes yields higher precedence,
as this implies that the literal itself is longer and thus more
specific, as literals may be followed by variable-length components.

This also helps reliable compute less-pathological cases, such as the
following:

- /foo/**/myfile*/**
- /foo/**/bar/myfile.txt

We may be able to tell that the latter has precedence, but a
deterministic algorithm to compute this is difficult to write and reason
about. By including a matching path, such as "/foo/bar/myfile.txt", we
are able to see that the "/**/" in the former pattern matches more
characters than that of the latter, and thus is a less precise match,
and has lower precedence.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>