i/prompting: do not error on unexpected apparmor permission
Rather than returning an error when there is no abstract permission
mapping to a received AppArmor file permission, instead log a message
with a warning and omit AppArmor permission from the abstract
permissions list.

Because the final reply notification is constructed from the
originally-requested permissions, rather than the AppArmor permissions
corresponding to the abstract permissions the user supplied, the ignored
unexpected permission will still be included in the reply.

If, in the future, this changes, further changes will be required to
ensure that requested permissions which are not mapped to an abstract
permission are still included in the reply notification message.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>