cmd/snapd-apparmor: add FIXME comments around flawed WSL logic
WSL-1 is an emulator running on top of Windows kernel. It doesn't include
the Linux kernel at all and has no support for Linux Security Modules or
LSMs. WSL-2 is a virtual machine running a variant of the Linux kernel
maintained by Microsoft, with individual distributions running inside
WSL containers (quasi containers).

At present the Microsoft kernel does not enable AppArmor and the runtime
environment does not enable AppArmor stacking, so from the point of view
of Snapd, apparmor is entirely unavailable.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>