cmd/snapd-apparmor: disable on WSL 1 and WSL 2
There's some confusing logic around what snapd-apparmor does on WSL-1
and WSL-2. Looking back at the historic commits I see that we first
tried to enable apparmor whenever WSL-2 is used and then somehow broke
the logic by differentiating WSL-1 and WSL-2 but using this knowledge
incorrectly:

- https://github.com/snapcore/snapd/pull/12179
- https://github.com/snapcore/snapd/pull/11349

With this I'm inclined to disable snapd-apparmor both on WSL-1 - where
it never made any sense as LSM stack is not emulated by Windows at all,
and on WSL-2 - where it is disabled in the Microsoft distribution of the
Linux kernel.

This leaves an unknown number of people with a non-Microsoft kernel on
WSL-2 that might plausibly use apparmor in a single distribution, as for
proper entablement stacking and initialization must happen in the
Microsoft init process that is beyond our control.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>