i/prompting/requestprompts: store prompts as slice in user prompt DB
Rather than using a map from ID to prompt, instead store all prompts for
the given user in a slice, where new prompts are appended to the end and
when a prompt is removed, the final prompt in the slice is moved to the
index of the removed slice, and the slice is truncated by one element.

To look up prompts by ID, the user prompt DB also stores a map from ID
to the index of the corresponding prompt in the prompt slice.

The benefit of this split storage system is that it is much easier to
extract the full list of prompts for a user, since they are already
stored as a slice. The one caveat is that we can't guarantee that the
slice won't be mutated internally after it has been returned, so we must
return a copy of the slice. The copy operation is still much faster than
iterating through all elements of a map and flattening it into a slice,
however.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>