interfaces: transition steam_support to use allow all,
Steam needs very broad permissions, and needs to be able to run a
broad set of different games with different permissions requirements.
Steam updates often require an update to the permissions in the
steam_support interface.

To ease support and ensure a good user experience move steam_support
to a profile that allows all permissions. This will prevent the steam
snap from breaking when steam updates, or a new game is added
requiring permissions that are not covered.

For newer versions of apparmor (4.0.2 and later) use the "allow all,"
rule, which will dynamically expand to all supported permissions.

For older versions of apparmor use the set of apparmor rule classes
supported before apparmor 4.0.2.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>