multiple: move split mount options logic to osutil
fstab entries with overlay options that could contain commas in the
path could not potentially be parsed properly. The logic that was
implementing that functionality was moved from a test case in
snap-bootstrap to mountentry_linux.go in osutil.

It was also identified that the moved code was causing one of the
tests to fail (TestParseMountEntry3). This bug was fixed.

A new test case was also added.