interfaces,sandbox: transition steam to permissive apparmor profile (#14229)
* sandbox/apparmor: mask allow all rule until apparmor 4.0.2

Compilation of allow all is broken in apparmor_parser until version
4.0.2 and later. Ubuntu 24.04 shipped with 4.0.0~beta3, and had an SRU
based on 4.0.1. Detect the version and mask the presence of allow all
unless apparmor parser 4.0.2, or newer, is used.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* sandbox/apparmor: remove stray newline

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* sandbox/apparmor: add feature probe for io_uring mediation

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces: transition steam_support to use allow all,

Steam needs very broad permissions, and needs to be able to run a
broad set of different games with different permissions requirements.
Steam updates often require an update to the permissions in the
steam_support interface.

To ease support and ensure a good user experience move steam_support
to a profile that allows all permissions. This will prevent the steam
snap from breaking when steam updates, or a new game is added
requiring permissions that are not covered.

For newer versions of apparmor (4.0.2 and later) use the "allow all,"
rule, which will dynamically expand to all supported permissions.

For older versions of apparmor use the set of apparmor rule classes
supported before apparmor 4.0.2.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces/builtin: apply userns, mqueue and io_uring conditionally

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Co-authored-by: John Johansen <john.johansen@canonical.com>