interfaces/lxd-support: add support for allow all in AppArmor
Similarly to the recent changes to the steam-support interface in #14229, the
lxd-support interface also needs to support an essentially unconfined LXD by
giving it all the possible permissions. This was previously done through the use
of the unconfined AppArmor profile mode, along with the userns,
permission. However, due to issues with apparmor_parser not correctly generating
all the required permissions when using the unconfined profile mode, it is
better to instead follow the same pattern as is now done in steam-support and
make use of the allow all, AppArmor parser feature/syntax when available, or to
fallback to generating a profile which includes all the possible permissions
when this is not available (as well as retaining the unconfined profile mode).

Signed-off-by: Alex Murray <alex.murray@canonical.com>