i/builtin: make registry role optional (#14222)
* i/builtin: make registry role optional

Roles are not intended to serve the purpose of access control (views
can already do that). The role only controls which of the snaps' hooks
are invoked when changing registry data. This commit makes the default
role "observer" in which case snapd calls a hook when a view changes.
The snap can optionally set role to manager to be called during the
alteration of the view.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* i/builtin: add short explanation; rm whitespace

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>