i/prompting/requestprompts: add dedicated json prompt constraints
Instead of mixing marshalled and ignored fields in the same
`PromptConstraints` struct, instead use an internal `promptConstraints`
type and a dedicated `promptConstraintsJSON` type for marshalling the
prompt as json. Since prompt constraints are always served by the API
and never received, there is no need for the ability to unmarshal json
into a `promptConstraints`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>