c/snap-bootstrap: rename FsLabel field to GptLabel
v1 CVM images use the filesystem label to auto-discover an encrypted
rootfs partition based on whether it has the '-enc' suffix. The new
images use the GPT label to auto-discover the verity partition. In
order to keep backwards compatibility, the customization tool will use
set both labels to the same one.

This commit clarifies that the field in the new structs actually contain
GPT label information.