c/snap-bootstrap: limit upper overlay folder creation to the tmpfs case
creating the upper and work folders used for setting up the overlayfs
mount, was initially handled by snap-bootstrap whether it set up a tmpfs
or detected a writable partition on disk but in order to support the
writable partition case, it was later discovered that these folders might
need to be created during the customization step and might already exist.

snap-bootstrap will now create those folders only if it has detected
no writable partition in the manifest and has auto-created a tmpfs-based
writable partition.