c/snap-bootstrap: revert comment for secbootProvisionForCVM
trying to boot an image on a system without a TPM present was failing
due to not-finding a TPM device. The source of the error was not the
provision step as secbotProvisionForCVM already ignores errors (returns
nil) if a TPM device is not found or a template file for provision the
TPM is not found.