i/p/requestprompts,s/a/n/listener: include explicit permissions in replies (#14196)
* i/p/requestprompts,s/a/n/listener: include AA permissions in replies

Include the AppArmor permissions allowed/denied by the user along with
the allow/deny decision in the response sent back to the listener. This
way, the listener can allow only the permissions which were explicitly
granted by the user, rather than automatically granting all requested
permissions.

One reason to do so is that if a request is received for several
AppArmor permissions and at least one of those permissions has no
mapping to an abstract permission for the interface in question, then
those unknown permissions can be discarded. If the request is eventually
allowed, then only the AppArmor permissions matching the abstract
permissions which the user explicitly approved are allowed, and the
unknown permissions are implicitly denied.

On the listener side, the listener receives two permissions bitmasks
from the kernel (for mediation class file): Allow and Deny.

If the user grants the requested permissions, then the allowed bits are
`(Allow | (Deny & approved))`, while the denied bits are `0`.

If the user denies the requested permissions, then the allowed bits are
`Allow` and the denied bits are `Deny` (same as what was received).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting,s/a/n/listener: adjust doc comments and replace interface{} with any

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting/requestprompts: add dedicated json prompt constraints

Instead of mixing marshalled and ignored fields in the same
`PromptConstraints` struct, instead use an internal `promptConstraints`
type and a dedicated `promptConstraintsJSON` type for marshalling the
prompt as json. Since prompt constraints are always served by the API
and never received, there is no need for the ability to unmarshal json
into a `promptConstraints`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting/requestprompts: remove json marshalling for prompts

In the future, when we add the prompting API endpoints and/or write
prompts to disk so they can be persisted across snapd restarts, we
should re-add marshalling as needed.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>