sandbox/apparmor: backport improved parser feature probe
This is not a direct cherry pick, as we used squash merges and I no longer
had the original history. This may conflict a little when merged back into
snapd mainline branch.

The improved parser feature can detect that a feature is only functioning
correctly starting with a specific version of the parser.

Test code is a lot more robust and efficient, avoiding the exponential
shell program problem.

The allow-all feature is now detected, as is io-uring.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>