cmd/snap-confine: add generator of nvidia globs
To keep only one set of globs, preferrably those from the nvidia/ package
where we have unit tests to ensure that they are correct, we need to
generate the globs that are currently consumed from snap-confine.

The generator program uses a text/template and an embedded file system
to store a C source-code template.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>