tests: add integration test for nvidia libs
The test installs the test-snapd-nvidia snap, installs Nvidia userspace
libraries on the host, pretends that the kernel modules is loaded and,
inside the snap, dlopens all the libraries exposed in
/var/lib/snapd/lib/gl and gl32.

The test makes sure that no unresolved symbols occurred, no confinement
issues occurred and that certain key libraries were indeed loaded. In a
way, it's a very practical test for the underlying essentials that does
not really depend on specific hardware and does not measure actual
display or rendered images.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>