i/p/requestrules: only prune expired rules on load or conflict
Rather than opportunistically pruning expired rules from the database
whenever they are encountered, instead only prune expired rules when the
database is first being populated, or when a new or patched rule
conflicts with an expired rule.

Since checking whether a rule is expired is fast, we can simply ignore
any expired rules when doing most operations on the database, thus
avoiding unnecessary database file writes as well as notices in many
cases where we're otherwise performing a read-only operation.

When adding or patching a rule, we're changing the database in a way
which requires saving it to disk anyway, and we can't have conflicting
pattern variants in the DB, so it's worth removing any expired rules
which conflict, and recording notices accordingly.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>