i/p/requestrules: add function to get variant map without mutating tree
Previously, `permissionDBForUserSnapInterfacePermission` created
intermediate data structures as it progressed from user DB to snap,
interface, and permission DB. This requires holding a write lock on the
database, but this function was often used in otherwise read-only
contexts.

Thus, this PR splits the function into two variants:
- `permissionDBForUserSnapInterfacePermission` returns the permission DB
  if it exists, along with a `bool` indicating whether it exists.
- `ensurePermissionDBForUserSnapInterfacePermission` does what
  `permissionDBForUserSnapInterfacePermission` had previously done,
  creating intermediate data structures.

When possible, callers should favor the former, as it only requires a
read lock instead of a write lock.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>