i/p/requestrules: reset db if any bad rules on disk
Instead of pruning conflicting permissions and invalid rules using
`refreshTreeEnforceConsistency`, simply discard the whole rule DB if
there is anything wrong with the rules on disk.

Rules stored to disk should only ever be saved from a rule DB, and since
rule DBs should always be internally consistent and valid, the rules on
disk should be valid and consistent as well. Thus, if they are not, they
must have been corrupted in some way, or otherwise the product of a bad
rule DB. Thus, it's safer to simply discard the DB and start over.

This also removes a lot of complexity.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>