i/p/requesrules: favor add/remove methods which maintain consistency
Rather than separately calling methods which add rules to the rules list
and the per-user rules tree of the database, respectively (or, WLOG,
removes), instead make the `addRule` and `removeRuleByID` methods act on
both these data structures consistently.

However, when removing rules, if an error occurs while saving the DB, we
need to re-add any removed rules to ensure no change occurred in the DB.
However, it is much riskier to remove and re-add a rule to the per-user
rules tree than it is to do so for the rules list. Thus, we don't want
to remove the rule from the rules tree until we know the save has
succeeded, thus guaranteeing there will be no need to roll back.

Therefore, we do preserve `addRuleToRulesList` and
`removeRuleByIDFromRulesList` as standalone helpers, with warnings that
the more consistent `addRule` and `removeRuleByID` should be used
instead whenever possible.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>