o/ifacestate: simplify interfaces requests manager
The former `apparmorprompting.Prompting` type has been renamed to be
`InterfacesRequestsManager`, and no longer follows the `Connect`, `Run`,
`Stop` interface.

Since `AppArmorPrompting` should be enabled or disabled on startup and
require a restart to change the value after that, there is no need to
use the `Ensure` method of the interface manager to start and stop the
requests manager, and there is no need to support starting and stopping
the manager after it has been created. Instead, it can be created and
immediately run as part of `StartUp`, and the `Stop` method can be
responsible for closing the listener and the prompts and rules backends.

If in the future we want to support dynamically enabling/disabling
prompting according to changes in the `"apparmor-prompting"` flag, its
configuration, or the presence of prompt handler clients, we may want to
move the checks for whether prompting is running into the `Ensure`
method, and use this to call `m.regenerateAllSecurityProfiles`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>