i/prompting: do not error on unrecognized aa permissions
Snapd may receive from the kernel AppArmor permissions which are not
recognized as mapping to an abstract permission for the interface in
question. This should not throw an error, and instead we should prompt
the user for any permissions which _are_ recognized, and only include
AppArmor permissions which correspond to those abstract permissions in
the response to the kernel.

By default, any permissions which are omitted from the "allow" field of
the message to the kernel will be auto-denied by the kernel, so this is
safe.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>