o/i/apparmorprompting: add partial unit tests and fix bugs
Most importantly, the uninitialized m.notifyPrompt and m.notifyRule were
being used when creating the prompts and rules backends, causing a
segmentation fault. That's fixed here.

Also, the listener and requestprompts packages had to export additional
struct fields and mock functions so that the apparmorprompting tests
could create and act on necessary structs.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>