o/i/apparmorprompting: add lock when synchronization required
When an operation requires synchronization between the prompts and rules
backends, such as when receiving a request from the kernel, adding a new
rule which must be checked against outstanding prompts, or replying to a
prompt with a lifespan other than "single", a lock must be held for
writing. However, when removing rules, or doing other operations which
only access one of the two backends at a time, only a read lock is
required.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>