o/registrystate: make registry transactions serializable
In order to attach registry transactions to tasks and changes, we need
to be able to serialize them. This change moves the transaction code
from registry/ to registrystate/ so we can replace the dependency on
unserializable callbacks to direct calls to update and read databags.
Similarly, this replaces references to a registry with its identifying
account and name pairs so we don't have to serialize an entire registry
(including schemas, rules, etc) each time we write a transaction.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>