o/i/apparmorprompting: clarify response on denied request
If there is an allow rule for some permissions and a deny rule for other
permissions which match a received request, a response should be sent
immediately which allows only the previously-allowed permissions. All
previously-denied permissions, as well as those which were not matched
by any rule, should be automatically denied by the kernel due to their
omission from the allow field in the response message.

The listener is responsible for mapping {allow/deny, perms} to an allow
and deny permission mask.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>