o/i/apparmorprompting: deny all requests for root user
If a request is received for root, deny it immediately. There is no user
daemon for root, so there should be prompting client able to handle it,
users should not generally be running snaps manually as root, and
services should not be using prompt-enabled interfaces. Thus, it is
safest to simply deny any requests for the root user.

Also, add various unit tests.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>