s/a/n/listener: handle permissions which aa both allowed and denied
If an operation was matched by both `allow` and `prompt` rules in the
apparmor profile, the message from the kernel may contain overlapping
permissions in the allow and deny fields. It is important that any
such permissions are not auto-allowed, and that we do what the user
specifies, either via existing request rules or a direct prompt reply.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>