many: replace listener response with allowed permission
When a prompt is denied, either because of a direct reply or because a
rule was added which denies one of its remaining permissions, the prompt
may have some originally-requested permissions which were previously
allowed due to previous rules. We want to reply to the kernel that those
previously-allowed permissions were allowed, but the newly-denied
permissions and any permissions which were not previously allowed by a
rule should be denied.

The old `listener.Response` type had a field for whether the request
should be allowed or denied, as well as the permissions for which that
decision should apply. Because of the cases described above, users of
this type often needed to send an "allow" response with the previously-
allowed permissions, even if the reply which triggered the response was
actually a denial. Furthermore, in the deny case, the permissions
included in the response were ignored, and all originally-requested
permissions were denied.

This is clearly confusing.

A much simpler approach is to reply directly with the permission(s)
which should be allowed. In strictly "deny" cases where no permissions
were previously allowed, the allowed permission sent as the response
should be the empty permission for the corresponding AppArmor permission
type -- in the case of a file, this is `notify.FilePermission(0)`. In
the more common cases where there may be some previously-allowed
permissions, any such permissions can be sent as the allowed permission,
and any permissions which are not included in the allowed permission are
denied by the listener.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

s/a/n/listener: accept nil allowed permission as allowing no permissions

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

s/a/n/listener: test response behavior with different allowed permission

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

s/a/n/listener: use the &^ operator instead of & ^ separately

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>